backend/requirements.txt
fastapi
uvicorn
manim
google-genai
pydantic
python-dotenv
edge-tts
openai
requests
